src/order-executor/requirements.txt
alpaca-py==0.38.0
confluent-kafka==2.8.0
fastapi==0.115.12
uvicorn==0.34.2
python-dotenv==1.1.0
structlog==25.1.0
pytz==2025.2
